Renamed the DomainOIDCService and improved code and comments.
